requirements.txt
astroid==2.3.2
certifi==2019.9.11
chardet==3.0.4
colorama==0.4.1
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.4
coveralls==1.8.2
dj-database-url==0.5.0
Django==2.2.6
django-cors-headers==3.1.1
django-heroku==0.3.1
djangorestframework==3.10.3
docopt==0.6.2
drf-yasg==1.17.0
gunicorn==19.9.0
idna==2.8
inflection==0.3.1
isort==4.3.21
itypes==1.1.0
Jinja2==2.10.3
lazy-object-proxy==1.4.2
MarkupSafe==1.1.1
mccabe==0.6.1
packaging==19.2
psycopg2==2.8.3
PyJWT==1.7.1
pylint==2.4.3
pyparsing==2.4.2
pytz==2019.3
PyYAML==5.1.2
requests==2.22.0
ruamel.yaml==0.16.5
ruamel.yaml.clib==0.2.0
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
uritemplate==3.0.0
urllib3==1.25.6
whitenoise==4.1.4
wrapt==1.11.2
